pyproject.toml
[project]
name = "qrize"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.1.0",
    "pyperclip>=1.9.0",
    "qrcode>=8.0",
    "reportlab>=4.3.1",
    "typer>=0.15.2",
]
